[#13] refactor: modularize walker and reduce duplication
